Extract voice input pipeline from App.jsx into useVoiceInput hook
Moves the full record -> transcribe -> AI-parse -> apply flow out of
App.jsx verbatim into src/hooks/useVoiceInput.js: the modal open/close
reset effect, MediaRecorder + native-Android recording, the
auto-start-on-shortcut effect, transcript parsing into new tasks and
edit commands, voiceApplyAllChanges (task creation + the edit-command
switch), and the voice-modal keyboard shortcuts.

The hook is called after useTaskActions because voiceApplyAllChanges
needs moveToRecycleBin from it; the voice*Ref indirections are kept so
voiceStopRecording's inline parse keeps reading fresh helpers at call
time. Voice state and refs stay owned by useVoiceAI. Also prunes the
six App.jsx imports that moved with the code.

This completes the App.jsx decomposition candidates from CLAUDE.md.
